pyproject.toml
[project]
name = "simulate-decision"
version = "0.1.0"
description = "Decision Simulator - AI-Powered Decision Analysis Engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy>=2.5.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=0.1.50",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "pyright>=1.1.350",
    "httpx>=0.27.0",
]

[project.scripts]
simulate-decision = "simulate_decision.cli.main:app"
simulate-decision-server = "simulate_decision.server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simulate_decision"]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
    "B",
    "C4",
    "SIM",
    "RUF",
]
ignore = [
    "E501",
    "SIM108",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingTypeStubs = "none"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
